Support for (PrimitiveTypeExpression).ToString function. - engine changes - more working samples - tested and works fine on postgre and sqlserver vendors.
Not complete for (NonPrimitiveTypeExpression).ToString: these cases should work when such expression can be executed in clr but should throw a NotSupportedException when it can't. With this commit a NonPrimitiveTypeExpression throws a NotSupportedException in both cases so it should be improved.
Two samples about these behaviors can be seen in: ReadTests_Conversions.cs/TestToString02 and TestToString04(currently fails)

git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@853 33146b6e-df29-0410-a1ca-39be37fb9416