Cleanup DataContext.GetVendor().
The IVendor lookup logic was failing under MONO_STRICT (i.e. nothing worked
when used as Mono's System.Data.Linq.dll), and the GetVendor() code was overly
complicated, had to much #if logic, and was doing work it didn't need to do
(such as using ConstructorInfo.Invoke() to create an object instead of just
using Activator.CreateInstance()).

Error messages were also cleaned up (i.e. it was looking for a default
constructor, but the error message said it was looking for a constructor with
a string parameter).

Unit tests for the DbLinqProvider and DbLinqConnectionType connection string
parameters were also added.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1012 33146b6e-df29-0410-a1ca-39be37fb9416