Remove Microsoft.VisualStudio.QualityTools.UnitTestFramework.dll.
The licensing of Microsoft.VisualStudio.QualityTools.UnitTestFramework.dll is
that it's a part of VisualStudio, and thus follows the normal VS licensing.
However, it's only included in VS Pro (maybe Standard, not sure, but certainly
not Express), so having a copy in svn is likely Bad.  It's also unnecessary,
since most of the tests use NUnit anyway.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1041 33146b6e-df29-0410-a1ca-39be37fb9416