Fix the ReadTests_Maths.Log tests.
There are many [Order Details] rows with a Discount value of 0, and Math.Log(0)
generates a floating point exception, causing the L2SQL tests to fail.

Fix the tests to add 1 to the Discount, thus giving a valid value to the
Math.Log() method.  This fixes the tests for L2SQL and SQL Server (and fixes
ReadTests_Maths.Log03() for SQLite).


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1149 33146b6e-df29-0410-a1ca-39be37fb9416