Fix regression in ReadTest.A6_ConnectionOpenTest() caused by r1154.
The problem is that ReadTest.A6_ConnectionOpenTest() tries to do:

	IDbConnection connection = ...;
	connection.Open();
	new Northwind(connection);

This was failing inside DataContext.Init(), as it was trying to always set
IDbConnection.ConnectionString (in case it was modified by the DbLinqProvider
removal code), which is taboo if the IDbConnection is already open.

We fix this by not setting IDbConnection.ConnectionString anymore (as it
shouldn't be needed).

We also need to change DataContextTest under L2SQL as DbLinqProvider/etc.
won't be removed by L2SQL (obviously), so the test should fail there.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1157 33146b6e-df29-0410-a1ca-39be37fb9416