Fix a KeyNotFoundException when running DbMetal.exe against Northwind.db3.
The problem is as follows: SQLite does no normalization.  Consequently, within
tests/instnwnd.SQLite.sql we have:

	CREATE TABLE [CustomerCustomerDemo] (
		... 
		REFERENCES [Customers] ([CustomerID])
	);

Thus, when we query SQLite for foreign-key information:

	PRAGMA foreign_key_list('CustomerCustomerDemo');

SQLite doesn't do any normalization, and dutifully returns:

	0|0|[Customers]|CustomerID|CustomerID|RESTRICT|RESTRICT|NONE

However, we find all tables via:

	SELECT tbl_name, 'main'
	FROM sqlite_master WHERE type='table' order by tbl_name;

which doesn't preserve any quoting (not that it matters, as in the CREATE
TABLE definition "Customers" is used instead of [Customers]).

The short of it is that later, when doing a dictionary lookup we try to lookup
the table "[Customers]", which fails with a KeyNotFoundException, as we should
actually be looking for the "Customers" table.

Go figure.

The fix is to remove all quoting when inserting table names, thus allowing
"Customers" to match "Customers", and everything is hunky-dory.

	* DbMetal/Test/CreateEntitiesFromSqliteDbTest.cs: Fix the unit tests
	  to (1) load Program into a new AppDomain, allowing us to (2) specify
	  the DbMetal.exe.config configuration file, which allows DbMetal to
	  actually find all the providers/etc. needed to do anything useful.
	  We also add an assertion that nothing is written to stderr, implying
	  that the program actually worked correctly.
	  Forthcoming: validation that the generated file is correct!
	* DbLinq.Sqlite/SqliteSchemaLoader.cs,
	  DbLinq.Sqlite/SqliteSchemaLoader.Constraints.cs: Fix the
	  aforementioned KeyNotFoundException; see above.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1255 33146b6e-df29-0410-a1ca-39be37fb9416