Remove GAC lookup support from DbMetal.
Attempting to lookup types from the GAC is pure, unadulterated, evil.  It's
evil firstly because the GAC is platform-specific, and can and will change.
(Consider .NET 2.0, which added x86- and x86-64-specific GAC's in addition to
the "normal" IL-GAC.)  It's also *extremely* brittle: by convention, API
compatibility is only guaranteed if the major version number stays the same.
Yet BY DESIGN, the DbMetal GAC lookup code was looking for the most recent
version number...which could be anything, including an incompatible API.

Then, it breaks things under Mono.

The correct fix is to use fully-qualified type names, OR to treat the ADO.NET
provider assemblies as private assemblies and place them with all your other
private assemblies (e.g. in the same directory as your .exe), which will allow
an assembly-qualified type name to find them even if a version is lacking.

This also adds a --debug option to DbMetal/SqlMetal.  By default stack traces
won't be shown, as they're ugly, and their length can make it difficult to see
the error message (as the stack trace is much longer than the error message).
If --debug isn't provided, we'll just show the exception message; if it is
provided, we'll also print the stack trace.

Fixes Issue 215.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1330 33146b6e-df29-0410-a1ca-39be37fb9416