CodeDom: Fix Northwind VB code compilation.
Previously, properties of e.g. array type would generate a property setter of:

	Public Property Picture() As Byte()
		Get
			Return Me._picture
		End Get
		Set
			If ((_picture <> value)) Then
				Me.OnPictureChanging(value)
				Me.SendPropertyChanging
				Me._picture = value
				Me.SendPropertyChanged("Picture")
				Me.OnPictureChanged
			End If
		End Set
	End Property

Alas, vbc complains about the above:

	error BC31080: Operator '<>' is not defined for types '1-dimensional array 
	of Byte' and '1-dimensional a rray of Byte'. Use 'Is' operator to compare 
	two reference types.

As mentioned, vbc wants us to use the Is operator.  Since we actually want 
*in*equality, we would want the IsNot operator:

	If ((_picture IsNot value)) Then

Unfortunately, CodeDom doesn't allow generation of the IsNot operator.  So we 
"hack" it by checking the property type.  If the property type is not a class
or nullable type, we continue with using '<>' (as was previously done).

If the property type IS a class or nullable type, we instead generate:

	If ((_picture Is value) = false) Then

This is necessary as CodeDom doesn't have support Unary negation, so we need
to compare against 'false' in order get the desired negation.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1359 33146b6e-df29-0410-a1ca-39be37fb9416